Restructure ModalView to JoinModal
