+++ Extension/Controller/EditProducto.php
@@ -45,7 +45,7 @@ class EditProducto
     public function createViewVariantLocations()
     {
         return function($viewName = 'ListVariantLocation') {
-            $this->addListView($viewName, 'ModelView\VariantLocation', 'locations', 'fas fa-search-location');
+            $this->addListView($viewName, 'JoinView\VariantLocation', 'locations', 'fas fa-search-location');
             $this->views[$viewName]->addOrderBy(['codewarehouse', 'aisle', 'rack', 'shelf', 'drawer'], 'warehouse');
             $this->views[$viewName]->addOrderBy(['aisle', 'rack', 'shelf', 'drawer', 'codewarehouse'], 'location');
             $this->views[$viewName]->searchFields = ['aisle', 'rack', 'shelf', 'drawer'];
